マッチングの場所情報から天気情報を取得することは可能にするまで。どうやら正確な位置情報は緯度と軽度から出した方が良さそうなので、そこの部分は位置情報が使えるようになってから。
