+++ app/views/matchings/edit.html.erb
@@ -3,6 +3,7 @@
 <!-- ユーザーの名前 -->
 <%= @user.user_profile.name %>
 <%= link_to 'プロフィール詳細', user_path(@user) %>
+
 <%= form_with(model: @matching, url: matching_path(@matching), method: :patch) do |f| %>
   <!-- マッチング名の編集 -->
   <div>
@@ -33,6 +34,14 @@
     <%= f.check_box :public_flag, checked: @matching.public_flag %>
     <%= f.label :public_flag, '外部公開' %>
   </div>
+
+  <% if @weather_data %>
+    <%= @weather_data.inspect %>
+    <p>Temperature: <%= @weather_data['temperature'] %>°C</p>
+    <p>Weather Description: <%= @weather_data['description'] %></p>
+  <% else %>
+    <p>No weather data available.</p>
+  <% end %>
 
   <!-- 編集ボタン -->
   <%= f.submit '編集' %>
