+++ app/views/matchings/show.html.erb
@@ -23,3 +23,11 @@ matchinds/show
   <label>場所</label>
   <p><%= @matching.place %></p>
 </div>
+
+<!-- マッチングの場所に基づく天気情報-->
+<% if @weather_data %>
+  <p>Temperature: <%= @weather_data.temperature %>°C</p>
+  <p>Weather Description: <%= @weather_data.description %></p>
+<% else %>
+  <p>No weather data available.</p>
+<% end %>
